Admin previliges for asset
assetid will be store irrespective of user account address. To confirm an asset owner already registered the asset the function modifier is used.

This enables us to check asset is already registered.  If asset is registered modifier restrict to store the asset by any or by its own owner.

From the point of view of security , spoofing can be  elliminated .